requirements.txt
google-api-python-client
google-cloud-core
google-cloud-speech
google-cloud-texttospeech
gtts
twitter
vlc
